+++ App/TCCEmptyApp/app/src/main/java/com/example/tcc/tccemptyapp/fragments/NewsFragment.java
@@ -7,10 +7,9 @@ import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.ProgressBar;
 
 import com.example.tcc.tccemptyapp.R;
-import com.example.tcc.tccemptyapp.adapters.NewsAdapter;
+import com.example.tcc.tccemptyapp.adapters.news.NewsAdapter;
 import com.example.tcc.tccemptyapp.helpers.ScrollHelper;
 import com.example.tcc.tccemptyapp.models.news.NewsList;
 import com.example.tcc.tccemptyapp.providers.news.NewsProvider;
